Refactor TraceAnnotator to separate utility class
