+++ supervision/annotators/core.py
@@ -1,15 +1,12 @@
 from math import sqrt
-from typing import List, Tuple, Union
+from typing import List, Tuple, Union, Optional
 
 import cv2
 import numpy as np
 
-from supervision.annotators.base import (
-    BaseAnnotator,
-    ColorMap,
-    resolve_color,
-    resolve_color_idx,
-)
+from supervision.annotators.base import BaseAnnotator
+from supervision.annotators.utils import ColorMap, resolve_color, resolve_color_idx, \
+    Trace
 from supervision.detection.core import Detections
 from supervision.draw.color import Color, ColorPalette
 from supervision.geometry.core import Position
@@ -561,4 +558,79 @@ class LabelAnnotator:
                 thickness=self.text_thickness,
                 lineType=cv2.LINE_AA,
             )
+        return scene
+
+
+class TraceAnnotator:
+    """
+    A class for drawing trace paths on an image based on detection coordinates.
+    """
+    def __init__(
+        self,
+        color: Union[Color, ColorPalette] = ColorPalette.default(),
+        position: Optional[Position] = Position.CENTER,
+        trace_length: int = 30,
+        thickness: int = 2,
+    ):
+        """
+        Args:
+            color (Union[Color, ColorPalette]): The color to draw the trace, can be
+                a single color or a color palette.
+            position (Optional[Position]): The position of the trace.
+                Defaults to `CENTER`.
+            trace_length (int): The maximum length of the trace in terms of historical
+                points. Defaults to `30`.
+            thickness (int): The thickness of the trace lines. Defaults to `2`.
+        """
+        self.color: Union[Color, ColorPalette] = color
+        self.position = position
+        self.trace = Trace(max_size=trace_length)
+        self.thickness = thickness
+
+    def annotate(self, scene: np.ndarray, detections: Detections) -> np.ndarray:
+        """
+        Draws trace paths on the frame based on the detection coordinates provided.
+
+        Args:
+            scene (np.ndarray): The image on which the traces will be drawn.
+            detections (Detections): The detections which include coordinates for
+                which the traces will be drawn.
+
+        Returns:
+            np.ndarray: The image with the trace paths drawn on it.
+
+        Example:
+            ```python
+            >>> import supervision as sv
+
+            >>> image = ...
+            >>> detections = sv.Detections(...)
+
+            >>> trace_annotator = sv.TraceAnnotator()
+            >>> annotated_frame = trace_annotator.annotate(
+            ...     scene=image.copy(),
+            ...     detections=detections
+            ... )
+            ```
+        """
+        self.trace.put(detections)
+
+        for i, (xyxy, mask, confidence, class_id, tracker_id) in enumerate(detections):
+            class_id = detections.class_id[i] if class_id is not None else None
+            idx = class_id if class_id is not None else i
+            color = (
+                self.color.by_idx(idx)
+                if isinstance(self.color, ColorPalette)
+                else self.color
+            )
+
+            xy = self.trace.get(tracker_id=tracker_id)
+            if len(xy) > 1:
+                scene = cv2.polylines(
+                    scene,
+                    [xy.astype(np.int32)],
+                    False,
+                    color=color.as_bgr(),
+                    thickness=self.thickness,
+                )
         return scene
